// src/shared/types/resume/contact.ts
export type Contact = {
  data: {
    email?: string
    phone?: string
    website?: string
    linkedin?: string
    github?: string
    twitter?: string
    facebook?: string
    instagram?: string
    youtube?: string
    pinterest?: string
    tiktok?: string
    snapchat?: string
    whatsapp?: string
    telegram?: string
    discord?: string
    skype?: string
    twitch?: string
    vimeo?: string
    reddit?: string
    quora?: string
    medium?: string
    stackoverflow?: string
    dribbble?: string
  }
}